Migrate GBEK tests with gcp secrets.
