+++ sdks/java/core/src/test/java/org/apache/beam/sdk/transforms/GroupByKeyTest.java
@@ -26,18 +26,12 @@ import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.collection.IsIterableContainingInAnyOrder.containsInAnyOrder;
 import static org.junit.Assert.assertThrows;
 
-import com.google.cloud.secretmanager.v1.ProjectName;
-import com.google.cloud.secretmanager.v1.SecretManagerServiceClient;
-import com.google.cloud.secretmanager.v1.SecretName;
-import com.google.cloud.secretmanager.v1.SecretPayload;
-import com.google.protobuf.ByteString;
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.Serializable;
-import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -96,9 +90,7 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 import org.hamcrest.Matcher;
 import org.joda.time.Duration;
 import org.joda.time.Instant;
-import org.junit.AfterClass;
 import org.junit.Assert;
-import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
@@ -621,110 +613,6 @@ public class GroupByKeyTest implements Serializable {
     @Category({ValidatesRunner.class, LargeKeys.Above100MB.class})
     public void testLargeKeys100MB() throws Exception {
       runLargeKeysTest(p, 100 << 20);
-    }
-  }
-
-  /** Tests validating GroupByKey behaviors with the gbek flag set. */
-  @RunWith(JUnit4.class)
-  public static class GbekTests extends SharedTestBase {
-    private static final String PROJECT_ID = "apache-beam-testing";
-    private static final String SECRET_ID = "gbek-test";
-    public static String gcpSecretVersionName;
-    private static String secretId;
-
-    @BeforeClass
-    public static void setup() throws IOException {
-      secretId = String.format("%s-%d", SECRET_ID, new SecureRandom().nextInt(10000));
-      SecretManagerServiceClient client;
-      try {
-        client = SecretManagerServiceClient.create();
-      } catch (IOException e) {
-        gcpSecretVersionName = null;
-        return;
-      }
-      ProjectName projectName = ProjectName.of(PROJECT_ID);
-      SecretName secretName = SecretName.of(PROJECT_ID, secretId);
-
-      try {
-        client.getSecret(secretName);
-      } catch (Exception e) {
-        com.google.cloud.secretmanager.v1.Secret secret =
-            com.google.cloud.secretmanager.v1.Secret.newBuilder()
-                .setReplication(
-                    com.google.cloud.secretmanager.v1.Replication.newBuilder()
-                        .setAutomatic(
-                            com.google.cloud.secretmanager.v1.Replication.Automatic.newBuilder()
-                                .build())
-                        .build())
-                .build();
-        client.createSecret(projectName, secretId, secret);
-        byte[] secretBytes = new byte[32];
-        new SecureRandom().nextBytes(secretBytes);
-        client.addSecretVersion(
-            secretName,
-            SecretPayload.newBuilder()
-                .setData(ByteString.copyFrom(java.util.Base64.getUrlEncoder().encode(secretBytes)))
-                .build());
-      }
-      gcpSecretVersionName = secretName.toString() + "/versions/latest";
-    }
-
-    @AfterClass
-    public static void tearDown() throws IOException {
-      if (gcpSecretVersionName != null) {
-        SecretManagerServiceClient client = SecretManagerServiceClient.create();
-        SecretName secretName = SecretName.of(PROJECT_ID, secretId);
-        client.deleteSecret(secretName);
-      }
-    }
-
-    @Test
-    @Category(NeedsRunner.class)
-    public void testGroupByKeyWithValidGcpSecretOption() {
-      if (gcpSecretVersionName == null) {
-        // Skip test if we couldn't set up secret manager
-        return;
-      }
-      List<KV<String, Integer>> ungroupedPairs =
-          Arrays.asList(
-              KV.of("k1", 3),
-              KV.of("k5", Integer.MAX_VALUE),
-              KV.of("k5", Integer.MIN_VALUE),
-              KV.of("k2", 66),
-              KV.of("k1", 4),
-              KV.of("k2", -33),
-              KV.of("k3", 0));
-
-      PCollection<KV<String, Integer>> input =
-          p.apply(
-              Create.of(ungroupedPairs)
-                  .withCoder(KvCoder.of(StringUtf8Coder.of(), BigEndianIntegerCoder.of())));
-
-      p.getOptions().setGbek(String.format("type:gcpsecret;version_name:%s", gcpSecretVersionName));
-      PCollection<KV<String, Iterable<Integer>>> output = input.apply(GroupByKey.create());
-
-      SerializableFunction<Iterable<KV<String, Iterable<Integer>>>, Void> checker =
-          containsKvs(
-              kv("k1", 3, 4),
-              kv("k5", Integer.MIN_VALUE, Integer.MAX_VALUE),
-              kv("k2", 66, -33),
-              kv("k3", 0));
-      PAssert.that(output).satisfies(checker);
-      PAssert.that(output).inWindow(GlobalWindow.INSTANCE).satisfies(checker);
-
-      p.run();
-    }
-
-    @Test
-    @Category(NeedsRunner.class)
-    public void testGroupByKeyWithInvalidGcpSecretOption() {
-      if (gcpSecretVersionName == null) {
-        // Skip test if we couldn't set up secret manager
-        return;
-      }
-      p.getOptions().setGbek("type:gcpsecret;version_name:bad_path/versions/latest");
-      p.apply(Create.of(KV.of("k1", 1))).apply(GroupByKey.create());
-      assertThrows(RuntimeException.class, () -> p.run());
     }
   }
